db column names removed _
