+++ src/main/java/com/parking/board/controller/BoardController.java
@@ -1,0 +1,8 @@
+package com.parking.board.controller;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class BoardController {
+
+}
