+++ src/main/java/com/parking/member/controller/MemberController.java
@@ -1,0 +1,6 @@
+package com.parking.member.controller;
+
+
+public class MemberController {
+
+}
