+++ src/main/java/sql/admin/admin-query.properties
@@ -1,4 +1,4 @@
 selectCountMember=select count(*) as cnt from member
-selectListPage=select * from (select rownum as rnum, a.* from (select * from member order by user_created_date desc) a) where rnum between ? and ?
+selectListPage=select * from (select rownum as rnum, a.* from (select * from member order by usercreateddate desc) a) where rnum between ? and ?
 #selectCountMemberPartial=select count(*) as cnt from member where ? like ?
-#selectListPagePartial=select * from (select rownum as rnum, a.* from (select * from member where ? like ? order by user_created_date desc) a) where rnum between ? and ?
+#selectListPagePartial=select * from (select rownum as rnum, a.* from (select * from member where ? like ? order by usercreateddate desc) a) where rnum between ? and ?
