+++ src/main/java/sql/api/api-query.properties
@@ -1,8 +1,8 @@
-insertparking_seoul=insert into PARKING_SEOUL values(?,?,?,?,?)
-selectListView=select * from parking_seoul where ps_addr like ?
-selectParkingSlot=select po.owner_business_no,po.owner_parking_code,ps.slot_user_code,ps.slot_begin_time,ps.slot_end_time from parking_owner po right join parking_slot ps on po.owner_business_no = ps.slot_business_no
-insertparking_slot=insert into parking_slot values(?,?,?,?)
+insertparking_seoul=insert into PARKINGSEOUL values(?,?,?,?,?)
+selectListView=select * from parkingseoul where psaddr like ?
+selectParkingSlot=select po.ownerbusinessno,po.ownerparkingcode,ps.slotusercode,ps.slotbegintime,ps.slotendtime from parking_owner po right join parking_slot ps on po.ownerbusinessno = ps.slotbusinessno
+insertparkingslot=insert into parkingslot values(?,?,?,?)
 insertParkingUserHistory=insert into userhistory values(default,?,?,default,?)
-insertparkingOwner=insert into parking_owner values(?,?)
+insertparkingOwner=insert into parkingowner values(?,?)
 insertCoupon=insert into coupon values(?,?,?,?,?)
 selectCouponList=select * from coupon
