+++ src/main/java/sql/board/qnaboard-query.properties
@@ -1,8 +1,8 @@
-selectQnaBoardList=select * from (select rownum as rnum, a.* from (select * from qnaboard order by qna_no desc, qna_created_date desc) a) where rnum between ? and ?
+selectQnaBoardList=select * from (select rownum as rnum, a.* from (select * from qnaboard order by qnano desc, qnacreateddate desc) a) where rnum between ? and ?
 
-selectQnaBoard=select * from qnaboard where qna_no=?
+selectQnaBoard=select * from qnaboard where qnano=?
 selectCountQnaBoard=select count(*) as cnt from qnaboard
 insertQnaboard=insert into qnaboard values(DEFAULT,?,?,?,DEFAULT,DEFAULT)
 selectCountQnaBoard=select count(*) as cnt from qnaboard
-selectPopularBoard=select * from qnaboard order by qna_readcount desc
+selectPopularBoard=select * from qnaboard order by qnareadcount desc
