+++ src/main/java/sql/board/review-query.properties
@@ -1,4 +1,4 @@
-selectReview=select * from review where review_no=?
-selectReviewList=select * from (select rownum as rnum, a.* from (select * from review order by review_no desc, review_created_date desc) a) where rnum between ? and ?
+selectReview=select * from review where reviewno=?
+selectReviewList=select * from (select rownum as rnum, a.* from (select * from review order by reviewno desc, reviewcreateddate desc) a) where rnum between ? and ?
 selectCountReview=select count(*) as cnt from review
 insertReview=insert into review values(DEFAULT,?,?,?,DEFAULT,?,DEFAULT)
