+++ src/main/java/sql/bookmark/bookmark-query.properties
@@ -1,5 +1,5 @@
-selectBookmark=select b.* from bookmark b JOIN member m ON b.bookmark_user_code = m.user_code and m.user_code=? and b.bookmark_parking_code=?
-selectBookmarkList=select * from bookmark b where b.bookmark_user_code=?
-selectBookmarkParkingList=select ps.* from bookmark b JOIN parking_seoul ps ON b.bookmark_parking_code = ps_parking_code where b.bookmark_user_code=?
+selectBookmark=select b.* from bookmark b JOIN member m ON b.bookmarkusercode = m.usercode and m.usercode=? and b.bookmarkparkingcode=?
+selectBookmarkList=select * from bookmark b where b.bookmarkusercode=?
+selectBookmarkParkingList=select ps.* from bookmark b JOIN parkingseoul ps ON b.bookmarkparkingcode = psparkingcode where b.bookmarkusercode=?
 insertBookmark=insert into bookmark values(DEFAULT,?,?)
-deleteBookmark=delete from bookmark where bookmark_user_code=? and bookmark_parking_code=?
+deleteBookmark=delete from bookmark where bookmarkusercode=? and bookmarkparkingcode=?
