+++ src/main/java/sql/history/userhistory-query.properties
@@ -1,5 +1,5 @@
-selectUserHistory=select h.* from userhistory h join parking_seoul ps on h.userhistory_parking_code = ps.ps_parking_code where h.userhistory_user_code=? and ps.ps_parking_code=?
-selectUserHistoryNo=select * from userhistory where userhistory_no=?
-selectParking=select ps.* from userhistory h join parking_seoul ps on h.userhistory_parking_code = ps.ps_parking_code where h.userhistory_user_code=? and ps.ps_parking_code=?
-selectUserHistoryList=select * from userHistory where userhistory_user_code=?
-selectUserHistoryParkingList=select ps.* from userhistory h join parking_seoul ps on h.userhistory_parking_code = ps.ps_parking_code where h.userhistory_user_code=?
+selectUserHistory=select h.* from userhistory h join parkingseoul ps on h.userhistoryparkingcode = ps.psparkingcode where h.userhistoryusercode=? and ps.psparkingcode=?
+selectUserHistoryNo=select * from userhistory where userhistoryno=?
+selectParking=select ps.* from userhistory h join parkingseoul ps on h.userhistoryparkingcode = ps.psparkingcode where h.userhistoryusercode=? and ps.psparkingcode=?
+selectUserHistoryList=select * from userHistory where userhistoryusercode=?
+selectUserHistoryParkingList=select ps.* from userhistory h join parkingseoul ps on h.userhistoryparkingcode = ps.psparkingcode where h.userhistoryusercode=?
