+++ src/main/java/sql/member/member-query.properties
@@ -1,9 +1,9 @@
-selectEmailPw=select * from member where user_email=? and user_pw=?
-selectEmail=select * from member where user_email=?
+selectEmailPw=select * from member where useremail=? and userpw=?
+selectEmail=select * from member where useremail=?
 insertMember=insert into member values(?,?,?,?,?,?,?,?,?,?,?,?,?,?)
-selectUserCode=select * from member where user_code=?
-updateLoginDate=update member set user_login_date=? where user_email=?
-updateMember=update member set user_phone=?, user_name=?, user_addr=?, user_sms_yn=?, user_email_yn=?, user_original_filename=?, user_renamed_filename=? where user_code=?
-deleteMember=delete from member where user_code=?
-activateMember = update member set user_email_verified=1 where user_email=?
-changepassword= update member set user_pw=? where user_email=?
+selectUserCode=select * from member where usercode=?
+updateLoginDate=update member set userlogindate=? where useremail=?
+updateMember=update member set userphone=?, username=?, useraddr=?, usersmsyn=?, useremailyn=?, useroriginalfilename=?, userrenamedfilename=? where usercode=?
+deleteMember=delete from member where usercode=?
+activateMember = update member set useremailverified=1 where useremail=?
+changepassword= update member set userpw=? where useremail=?
